Revamp back-end error handling for appointments and users
